+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## TBD
+## 9.0.0 (2024-06-10)
 
 ### Enhancements